summarizer/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.16,<3.17
django-cors-headers>=4.7,<4.8

transformers>=4.56,<5.0
torch

nltk>=3.9,<4.0
langdetect>=1.0.9,<2.0
yake>=0.4.8,<0.5
textblob>=0.19,<0.20

PyMuPDF>=1.26,<1.27

